Move cwd to header
