+++ packages/cli/src/ui/App.tsx
@@ -903,7 +903,12 @@ const App = ({ config, settings, startupWarnings = [], version }: AppProps) => {
           items={[
             <Box flexDirection="column" key="header">
               {!settings.merged.hideBanner && (
-                <Header version={version} nightly={nightly} />
+                <Header
+                  version={version}
+                  nightly={nightly}
+                  targetDir={config.getTargetDir()}
+                  branchName={branchName}
+                />
               )}
               {!settings.merged.hideTips && <Tips config={config} />}
             </Box>,
@@ -1195,9 +1200,7 @@ const App = ({ config, settings, startupWarnings = [], version }: AppProps) => {
           )}
           <Footer
             model={currentModel}
-            targetDir={config.getTargetDir()}
             debugMode={config.getDebugMode()}
-            branchName={branchName}
             debugMessage={debugMessage}
             corgiMode={corgiMode}
             errorCount={errorCount}
@@ -1206,7 +1209,6 @@ const App = ({ config, settings, startupWarnings = [], version }: AppProps) => {
               config.getDebugMode() || settings.merged.showMemoryUsage || false
             }
             promptTokenCount={sessionStats.lastPromptTokenCount}
-            nightly={nightly}
             vimMode={vimModeEnabled ? vimMode : undefined}
             isTrustedFolder={isTrustedFolderState}
           />
