+++ packages/cli/src/ui/components/Footer.tsx
@@ -7,11 +7,10 @@
 import React from 'react';
 import { Box, Text } from 'ink';
 import { theme } from '../semantic-colors.js';
-import { shortenPath, tildeifyPath } from '@google/gemini-cli-core';
+
 import { ConsoleSummaryDisplay } from './ConsoleSummaryDisplay.js';
 import process from 'node:process';
-import path from 'node:path';
-import Gradient from 'ink-gradient';
+
 import { MemoryUsageDisplay } from './MemoryUsageDisplay.js';
 import { ContextUsageDisplay } from './ContextUsageDisplay.js';
 import { DebugProfiler } from './DebugProfiler.js';
@@ -21,8 +20,6 @@ import { isNarrowWidth } from '../utils/isNarrowWidth.js';
 
 interface FooterProps {
   model: string;
-  targetDir: string;
-  branchName?: string;
   debugMode: boolean;
   debugMessage: string;
   corgiMode: boolean;
@@ -30,15 +27,12 @@ interface FooterProps {
   showErrorDetails: boolean;
   showMemoryUsage?: boolean;
   promptTokenCount: number;
-  nightly: boolean;
   vimMode?: string;
   isTrustedFolder?: boolean;
 }
 
 export const Footer: React.FC<FooterProps> = ({
   model,
-  targetDir,
-  branchName,
   debugMode,
   debugMessage,
   corgiMode,
@@ -46,19 +40,12 @@ export const Footer: React.FC<FooterProps> = ({
   showErrorDetails,
   showMemoryUsage,
   promptTokenCount,
-  nightly,
   vimMode,
   isTrustedFolder,
 }) => {
   const { columns: terminalWidth } = useTerminalSize();
 
   const isNarrow = isNarrowWidth(terminalWidth);
-
-  // Adjust path length based on terminal width
-  const pathLength = Math.max(20, Math.floor(terminalWidth * 0.4));
-  const displayPath = isNarrow
-    ? path.basename(tildeifyPath(targetDir))
-    : shortenPath(tildeifyPath(targetDir), pathLength);
 
   return (
     <Box
@@ -70,21 +57,6 @@ export const Footer: React.FC<FooterProps> = ({
       <Box>
         {debugMode && <DebugProfiler />}
         {vimMode && <Text color={theme.text.secondary}>[{vimMode}] </Text>}
-        {nightly ? (
-          <Gradient colors={theme.ui.gradient}>
-            <Text>
-              {displayPath}
-              {branchName && <Text> ({branchName}*)</Text>}
-            </Text>
-          </Gradient>
-        ) : (
-          <Text color={theme.text.link}>
-            {displayPath}
-            {branchName && (
-              <Text color={theme.text.secondary}> ({branchName}*)</Text>
-            )}
-          </Text>
-        )}
         {debugMode && (
           <Text color={theme.status.error}>
             {' ' + (debugMessage || '--debug')}
