+++ packages/cli/src/ui/components/Header.test.tsx
@@ -20,25 +20,65 @@ describe('<Header />', () => {
       columns: 120,
       rows: 20,
     });
-    const { lastFrame } = render(<Header version="1.0.0" nightly={false} />);
+    const { lastFrame } = render(
+      <Header
+        version="1.0.0"
+        nightly={false}
+        targetDir="/test/dir"
+        branchName="test-branch"
+      />,
+    );
     expect(lastFrame()).toContain(longAsciiLogo);
   });
 
   it('renders custom ASCII art when provided', () => {
     const customArt = 'CUSTOM ART';
     const { lastFrame } = render(
-      <Header version="1.0.0" nightly={false} customAsciiArt={customArt} />,
+      <Header
+        version="1.0.0"
+        nightly={false}
+        customAsciiArt={customArt}
+        targetDir="/test/dir"
+        branchName="test-branch"
+      />,
     );
     expect(lastFrame()).toContain(customArt);
   });
 
   it('displays the version number when nightly is true', () => {
-    const { lastFrame } = render(<Header version="1.0.0" nightly={true} />);
+    const { lastFrame } = render(
+      <Header
+        version="1.0.0"
+        nightly={true}
+        targetDir="/test/dir"
+        branchName="test-branch"
+      />,
+    );
     expect(lastFrame()).toContain('v1.0.0');
   });
 
   it('does not display the version number when nightly is false', () => {
-    const { lastFrame } = render(<Header version="1.0.0" nightly={false} />);
+    const { lastFrame } = render(
+      <Header
+        version="1.0.0"
+        nightly={false}
+        targetDir="/test/dir"
+        branchName="test-branch"
+      />,
+    );
     expect(lastFrame()).not.toContain('v1.0.0');
+  });
+
+  it('displays the target directory and branch name', () => {
+    const { lastFrame } = render(
+      <Header
+        version="1.0.0"
+        nightly={false}
+        targetDir="/test/dir"
+        branchName="test-branch"
+      />,
+    );
+    expect(lastFrame()).toContain('/test/dir');
+    expect(lastFrame()).toContain('(test-branch*)');
   });
 });
