+++ packages/cli/src/ui/components/Header.tsx
@@ -7,21 +7,29 @@
 import React from 'react';
 import { Box, Text } from 'ink';
 import Gradient from 'ink-gradient';
-import { Colors } from '../colors.js';
+
 import { shortAsciiLogo, longAsciiLogo, tinyAsciiLogo } from './AsciiArt.js';
 import { getAsciiArtWidth } from '../utils/textUtils.js';
+import { theme } from '../semantic-colors.js';
+import { shortenPath, tildeifyPath } from '@google/gemini-cli-core';
 import { useTerminalSize } from '../hooks/useTerminalSize.js';
+import { isNarrowWidth } from '../utils/isNarrowWidth.js';
+import path from 'node:path';
 
 interface HeaderProps {
   customAsciiArt?: string; // For user-defined ASCII art
   version: string;
   nightly: boolean;
+  targetDir: string;
+  branchName?: string;
 }
 
 export const Header: React.FC<HeaderProps> = ({
   customAsciiArt,
   version,
   nightly,
+  targetDir,
+  branchName,
 }) => {
   const { columns: terminalWidth } = useTerminalSize();
   let displayTitle;
@@ -39,6 +47,11 @@ export const Header: React.FC<HeaderProps> = ({
   }
 
   const artWidth = getAsciiArtWidth(displayTitle);
+  const isNarrow = isNarrowWidth(terminalWidth);
+  const pathLength = Math.max(20, Math.floor(terminalWidth * 0.4));
+  const displayPath = isNarrow
+    ? path.basename(tildeifyPath(targetDir))
+    : shortenPath(tildeifyPath(targetDir), pathLength);
 
   return (
     <Box
@@ -47,24 +60,48 @@ export const Header: React.FC<HeaderProps> = ({
       flexShrink={0}
       flexDirection="column"
     >
-      {Colors.GradientColors ? (
-        <Gradient colors={Colors.GradientColors}>
+      {theme.ui.gradient ? (
+        <Gradient colors={theme.ui.gradient}>
           <Text>{displayTitle}</Text>
         </Gradient>
       ) : (
         <Text>{displayTitle}</Text>
       )}
-      {nightly && (
-        <Box width="100%" flexDirection="row" justifyContent="flex-end">
-          {Colors.GradientColors ? (
-            <Gradient colors={Colors.GradientColors}>
-              <Text>v{version}</Text>
+      <Box
+        justifyContent="space-between"
+        width="100%"
+        flexDirection={isNarrow ? 'column' : 'row'}
+        alignItems={isNarrow ? 'flex-start' : 'center'}
+      >
+        <Box>
+          {nightly ? (
+            <Gradient colors={theme.ui.gradient}>
+              <Text>
+                {displayPath}
+                {branchName && <Text> ({branchName}*)</Text>}
+              </Text>
             </Gradient>
           ) : (
-            <Text>v{version}</Text>
+            <Text color={theme.text.link}>
+              {displayPath}
+              {branchName && (
+                <Text color={theme.text.secondary}> ({branchName}*)</Text>
+              )}
+            </Text>
           )}
         </Box>
-      )}
+        {nightly && (
+          <Box width="100%" flexDirection="row" justifyContent="flex-end">
+            {theme.ui.gradient ? (
+              <Gradient colors={theme.ui.gradient}>
+                <Text>v{version}</Text>
+              </Gradient>
+            ) : (
+              <Text>v{version}</Text>
+            )}
+          </Box>
+        )}
+      </Box>
     </Box>
   );
 };
